fix: namespace script, executable, and tmp dir
A single Buildkite node may run multiple agents, each of which may
run multiple steps (from different pipelines). Namespace all temp
files created by boost with the running buildkite step's uuid in order
to remove conflicts from concurrent steps.

This should resolve an issue where the CLI sometimes fails to install
(likely because a concurrent step is using the file already).

```log
[INFO] installing cli to /tmp/boost.sh
/var/lib/buildkite-agent/plugins/github-com-peaudecastor-boost-security-scanner-buildkite-plugin-v2-0/hooks/../lib/action.bash: /tmp/boost.sh: /bin/sh: bad interpreter: Text file busy
```

In the future, it would be nice to configure the plugin to install
the CLI only if it does not exist or if the existing one is an old
version (and not remove the cli after running the step). This would
speed up most builds, but would likely require a new
api.boostsecurity.io endpoint which provides the latest CLI version.